+++ src/psd_tools/api/mask.py
@@ -140,7 +140,7 @@ class Mask(MaskProtocol):
     @disabled.setter
     def disabled(self, value: bool) -> None:
         self._data.flags.mask_disabled = value
-        self._layer._psd._mark_updated()
+        self._layer._psd.mark_updated()
 
     @property
     def flags(self) -> MaskFlags: